compsci235-assignment2-really-cool-team/requirements.txt
attrs==21.2.0
click==7.1.2
Flask==1.1.4
Flask-WTF==0.15.1
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==2.0.1
packaging==21.0
pluggy==1.0.0
py==1.10.0
pyparsing==2.4.7
pytest==6.2.5
python-dotenv==0.19.0
toml==0.10.2
Werkzeug==1.0.1
WTForms==2.3.3
SQLAlchemy==1.4.23